Use sessionAffinity: ClientIP
This ensures that a client with a given IP address gets served by the
same container for the duration of a "session".

If we set sessionAffinity: None, downloading reports fails with number
of replicas > 1. This is due to the fact that the report is generated
on one container, while the URL to download it gets redirected by the
load balancer to another one, which has no knowledge of the generated
report.

See https://kubernetes.io/docs/concepts/services-networking/service/
for further details.